refactor: UserInfo 삭제
