+++ src/main/java/qna/domain/UserRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByUserInfo(UserInfo userInfo);
+    Optional<User> findByUserLogin(UserLogin userLogin);
 }
