+++ src/test/java/qna/jpa_test/WhereTest.java
@@ -32,8 +32,7 @@ public class WhereTest {
     @BeforeEach
     void setting() {
         UserLogin userLogin = new UserLogin("id", "pwd", "writer@slipp.net");
-        UserInfo userInfo = new UserInfo("writer", userLogin);
-        writer = userRepository.save(new User(userInfo));
+        writer = userRepository.save(new User("writer", userLogin));
         question = questionRepository.save(QuestionTest.Q1.writeBy(writer));
     }
